+++ android/src/main/kotlin/com/inappstory/inappstory_plugin/callbacks/CallToActionCallbackAdaptor.kt
@@ -27,7 +27,7 @@ class CallToActionCallbackAdaptor(
 
             flutterPluginBinding.runOnMainThread {
                 api.callToAction(
-                    slideDataArg = content?.let { mapSlideDataDto(it) },
+                    slideDataArg = content?.let { mapSlideDataDto(content) },
                     urlArg = url,
                     clickActionArg = action?.let { ClickActionDto.ofRaw(it.ordinal) }
                 ) {}