+++ android/src/main/kotlin/com/inappstory/inappstory_plugin/callbacks/SingleLoadCallbackAdaptor.kt
@@ -1,4 +1,4 @@
-package com.inappstory.inappstory_plugin.adaptors
+package com.inappstory.inappstory_plugin.callbacks
 
 import SingleLoadCallbackFlutterApi
 import com.inappstory.inappstory_plugin.mapStoryData
@@ -6,7 +6,6 @@ import com.inappstory.inappstory_plugin.runOnMainThread
 import com.inappstory.sdk.stories.outercallbacks.common.reader.StoryData
 import com.inappstory.sdk.stories.outercallbacks.common.single.SingleLoadCallback
 import io.flutter.embedding.engine.plugins.FlutterPlugin
-
 
 class SingleLoadCallbackAdaptor(
     private val flutterPluginBinding: FlutterPlugin.FlutterPluginBinding,