+++ ios/Classes/InAppStoryManagerAdaptor.swift
@@ -1,28 +1,34 @@
-import Foundation
 import Flutter
+import Foundation
 @_spi(IAS_API) import InAppStorySDK
 
 class InAppStoryManagerAdaptor: InAppStoryManagerHostApi {
     init(binaryMessenger: FlutterBinaryMessenger) {
-        InAppStoryManagerHostApiSetup.setUp(binaryMessenger: binaryMessenger, api: self)
+        InAppStoryManagerHostApiSetup.setUp(
+            binaryMessenger: binaryMessenger,
+            api: self
+        )
     }
-    
-    func setPlaceholders(newPlaceholders: [String : String]) throws {
+
+    func setPlaceholders(newPlaceholders: [String: String]) throws {
         InAppStory.shared.placeholders = newPlaceholders
     }
-    
+
     func setTags(tags: [String]) throws {
         InAppStory.shared.settings?.tags = tags
         InAppStory.shared.setTags(tags)
     }
-    
-    func changeUser(userId: String, completion: @escaping (Result<Void, Error>) -> Void) {
+
+    func changeUser(
+        userId: String,
+        completion: @escaping (Result<Void, Error>) -> Void
+    ) {
         InAppStory.shared.settings = Settings(userID: userId)
-        
+
         completion(.success(()))
     }
-    
+
     func closeReaders() throws {
-        InAppStory.shared.closeReader() {}
+        InAppStory.shared.closeReader {}
     }
 }