+++ ios/Classes/StoriesListUpdateHandlerAdaptor.swift
@@ -5,53 +5,72 @@
 //  Created by Sergey Salnikov on 29.10.2024.
 //
 
-import Foundation
 import Flutter
+import Foundation
 @_spi(IAS_API) import InAppStorySDK
 
 class StoriesListUpdateHandlerAdaptor {
-    init(binaryMessenger: FlutterBinaryMessenger, storyListAPI: StoryListAPI, uniqueId: String) {
+    init(
+        binaryMessenger: FlutterBinaryMessenger,
+        storyListAPI: StoryListAPI,
+        uniqueId: String
+    ) {
         self.binaryMessenger = binaryMessenger
-        
+
         self.uniqueId = uniqueId
-        
-        self.flutter = InAppStoryAPIListSubscriberFlutterApi(binaryMessenger: binaryMessenger, messageChannelSuffix: uniqueId)
-        
+
+        self.flutter = InAppStoryAPIListSubscriberFlutterApi(
+            binaryMessenger: binaryMessenger,
+            messageChannelSuffix: uniqueId
+        )
+
         self.storyListAPI = storyListAPI
-        
+
         storyListAPI.storyListUpdate = storiesListUpdateHandler
-        
+
         storyListAPI.storyUpdate = storyUpdateHandler
-        
+
         storyListAPI.favoritesUpdate = favoriteUpdateHandler
     }
-    
+
     private var binaryMessenger: FlutterBinaryMessenger
 
     private var uniqueId: String
-    
+
     private var storyListAPI: StoryListAPI
-    
+
     private var flutter: InAppStoryAPIListSubscriberFlutterApi
-    
-    private lazy var storiesListUpdateHandler: StoriesListUpdateHandler = { storiesList, isFavorite, feed in        
-        self.flutter.updateStoriesData(list: storiesList.map(self.mapStoryAPIData), completion: {_ in })
-        
+
+    private lazy var storiesListUpdateHandler: StoriesListUpdateHandler = {
+        storiesList,
+        isFavorite,
+        feed in
+        self.flutter.updateStoriesData(
+            list: storiesList.map(self.mapStoryAPIData),
+            completion: { _ in }
+        )
+
         self.favoriteUpdateHandler(isFavorite)
     }
-    
+
     private lazy var storyUpdateHandler: StoryUpdateHandler = { storyCellData in
-        self.flutter.updateStoryData(var1: self.mapStoryAPIData(arg: storyCellData), completion: {_ in })
+        self.flutter.updateStoryData(
+            var1: self.mapStoryAPIData(arg: storyCellData),
+            completion: { _ in }
+        )
     }
-    
+
     func favoriteUpdateHandler(_ data: [InAppStorySDK.SimpleFavoriteData]?) {
         if let favorites = data {
-            if (!favorites.isEmpty) {
-                self.flutter.updateFavoriteStoriesData(list: favorites.map(self.mapFavorite), completion: {_ in })
+            if !favorites.isEmpty {
+                self.flutter.updateFavoriteStoriesData(
+                    list: favorites.map(self.mapFavorite),
+                    completion: { _ in }
+                )
             }
         }
     }
-    
+
     private func mapStoryAPIData(arg: StoryCellData) -> StoryAPIDataDto {
         return StoryAPIDataDto(
             id: Int64(arg.storyID)!,
@@ -66,12 +85,12 @@ class StoriesListUpdateHandlerAdaptor {
             aspectRatio: Double(storyListAPI.cellRatio)
         )
     }
-    
+
     private func mapStoryData(arg: StoryData) -> StoryDataDto {
         return StoryDataDto(
             id: Int64(arg.id ?? "-1")!,
             title: arg.title,
-            tags: "", // TODO map array of tags to string
+            tags: "",  // TODO map array of tags to string
             feed: arg.feed,
             sourceType: mapStorySource(arg: arg.source),
             slidesCount: Int64(arg.slidesCount),
@@ -83,6 +102,8 @@ class StoriesListUpdateHandlerAdaptor {
         switch arg {
         case .story: return StoryTypeDto.cOMMON
         case .storyUGC: return StoryTypeDto.uGC
+        @unknown default:
+            return StoryTypeDto.cOMMON
         }
     }
 
@@ -92,11 +113,15 @@ class StoriesListUpdateHandlerAdaptor {
         case .list: return SourceTypeDto.lIST
         case .onboarding: return SourceTypeDto.oNBOARDING
         case .single: return SourceTypeDto.sINGLE
-    //    case FIXME NO stack: return SourceTypeDto.sTACK
+        @unknown default:
+            return SourceTypeDto.sINGLE
+        //    case FIXME NO stack: return SourceTypeDto.sTACK
         }
     }
-    
-    private func mapFavorite(arg: SimpleFavoriteData) -> StoryFavoriteItemAPIDataDto {
+
+    private func mapFavorite(arg: SimpleFavoriteData)
+        -> StoryFavoriteItemAPIDataDto
+    {
         return StoryFavoriteItemAPIDataDto(
             id: Int64(arg.serverID)!,
             imageFilePath: arg.image,